+++ server/src/main/java/luan/moonvs/controllers/ContentController.java
@@ -2,8 +2,10 @@ package luan.moonvs.controllers;
 
 import luan.moonvs.models.entities.Content;
 import luan.moonvs.models.enums.ContentType;
+import luan.moonvs.models.enums.SearchType;
 import luan.moonvs.models.responses.ContentSearch;
 import luan.moonvs.models.requests.ExternalContentRequest;
+import luan.moonvs.models.responses.Response;
 import luan.moonvs.services.ContentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -19,6 +21,9 @@ public class ContentController {
     @Autowired
     private ContentService service;
 
+    private final String HEADER_NAME = "message";
+
+    @Deprecated
     @GetMapping("/internal-search")
     public ResponseEntity<List<ContentSearch>> internalSearch(@RequestParam String title) {
         if (title == null || title.isBlank())
@@ -30,6 +35,7 @@ public class ContentController {
         return service.internalSearch(title);
     }
 
+    @Deprecated
     @GetMapping("/external-search")
     public ResponseEntity<List<ContentSearch>> externalSearch(@RequestParam String title) {
         title = title.replace(" ", "%20");
@@ -41,6 +47,24 @@ public class ContentController {
         return service.externalSearch(title);
     }
 
+    @GetMapping("/search")
+    public ResponseEntity<List<ContentSearch>> search(@RequestParam SearchType searchType, @RequestParam String title) {
+        final String EMPTY_SEARCH = "Please enter a content name!";
+
+        if (title == null || title.isBlank())
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .header(HEADER_NAME, EMPTY_SEARCH)
+                    .build();
+
+        var response = service.search(searchType, title);
+        return ResponseEntity
+                .status(response.status())
+                .header(HEADER_NAME, response.message())
+                .body(response.entity());
+    }
+
+    @Deprecated
     @GetMapping("internal/{id}")
     public ResponseEntity<Content> viewInternalContent(@PathVariable int id){
         if (id < 1)
@@ -51,6 +75,7 @@ public class ContentController {
         return service.internalContent(id);
     }
 
+    @Deprecated
     @GetMapping("/external")
     public ResponseEntity<?> viewExternalContent(@RequestParam int id, @RequestParam ContentType contentType) {
         if (id < 1)
@@ -59,5 +84,21 @@ public class ContentController {
                     .build();
 
         return service.externalContent(id, contentType);
+    }
+
+    @GetMapping("/view/{id}")
+    public ResponseEntity<Content> viewContent(@PathVariable int id, @RequestParam SearchType searchType, @RequestParam ContentType contentType) {
+        final String ZERO_OR_NEGATIVE_ID = "The requested content is out of range!";
+        if (id < 1)
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .header(HEADER_NAME, ZERO_OR_NEGATIVE_ID)
+                    .build();
+
+        var response = service.viewContent(id, searchType, contentType);
+        return ResponseEntity
+                .status(response.status())
+                .header(HEADER_NAME, response.message())
+                .body(response.entity());
     }
 }