+++ server/src/main/java/luan/moonvs/models/enums/SearchType.java
@@ -1,0 +1,12 @@
+package luan.moonvs.models.enums;
+
+public enum SearchType {
+    INTERNAL("internal"),
+    EXTERNAL("external");
+
+    private final String value;
+
+    SearchType(String value) {
+        this.value = value;
+    }
+}