+++ server/src/main/java/luan/moonvs/models/requests/UserAccountRequest.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.time.LocalDate;
 import java.util.UUID;
 
-@JsonIgnoreProperties
+@JsonIgnoreProperties(ignoreUnknown = true)
 public record UserAccountRequest(UUID idUser,
                                  String username,
                                  String email,