+++ server/src/main/java/luan/moonvs/security/SecurityConfig.java
@@ -34,7 +34,7 @@ public class SecurityConfig {
                 .csrf(AbstractHttpConfigurer::disable)
                 .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
                 .authorizeHttpRequests(authz -> authz
-                        .requestMatchers("/auth/", "/auth/login", "/auth/register", "auth/logout").permitAll()
+                        .requestMatchers("/auth/", "/auth/login", "/auth/register", "auth/logout", "/auth/sign-in", "/auth/sign-up").permitAll()
                         .anyRequest().authenticated()
                 )
                 .addFilterBefore(securityFilter, UsernamePasswordAuthenticationFilter.class)